feat: Change number_of_rooms from Integer to Float to allow half-rooms

+++ app/models/property.py
@@ -31,7 +31,7 @@ class Property(Base):
     purchase_price = Column(Float, nullable=False)
     image_url = Column(String, nullable=True)
     property_owner = Column(String, nullable=True)
-    number_of_rooms = Column(Integer, nullable=True)
+    number_of_rooms = Column(Float, nullable=True)
     parking_numbers = Column(Text, nullable=True)  # JSON array of strings
     electricity_meter_number = Column(String, nullable=True)
     electricity_account_number = Column(String, nullable=True)
